latexoperation()

latexoperation()

This function is not exported.

Latexify.latexoperation — Function.

latexoperation(ex::Expr, prevOp::AbstractArray)

Translate a simple operation given by ex to LaTeX maths syntax. This uses the information about the previous operations to deside if a parenthesis is needed.

source